+++ src/aux-records/RecordsServer.spec.ts
@@ -438,6 +438,9 @@ describe('RecordsServer', () => {
         generateLogoutUrl: jest.Mock<
             ReturnType<PrivoClientInterface['generateLogoutUrl']>
         >;
+        resendConsentRequest: jest.Mock<
+            ReturnType<PrivoClientInterface['resendConsentRequest']>
+        >;
     };
 
     beforeEach(async () => {
@@ -489,6 +492,7 @@ describe('RecordsServer', () => {
             checkEmail: jest.fn(),
             checkDisplayName: jest.fn(),
             generateLogoutUrl: jest.fn(),
+            resendConsentRequest: jest.fn(),
         };
         relyingParty = {
             id: 'relying_party_id',
@@ -503,6 +507,12 @@ describe('RecordsServer', () => {
             privoClient,
             [relyingParty]
         );
+
+        // manually disable the Privo flag for tests
+        // (it is automatically set to true because a privo client is specified, but most tests
+        // assume privo isn't enabled)
+        authController.privoEnabled = false;
+
         livekitController = new LivekitController(
             livekitApiKey,
             livekitSecretKey,
@@ -2800,6 +2810,8 @@ describe('RecordsServer', () => {
                 redirectUri: 'redirectUri',
                 ageOfConsent: 18,
             };
+
+            authController.privoEnabled = true;
         });
 
         it('should return a login request with the authorization URL', async () => {
@@ -2900,6 +2912,8 @@ describe('RecordsServer', () => {
                 redirectUri: 'redirectUri',
                 ageOfConsent: 18,
             };
+
+            authController.privoEnabled = true;
         });
 
         it('should save the authorization code', async () => {
@@ -3112,7 +3126,7 @@ describe('RecordsServer', () => {
                 idToken: 'idToken',
                 expiresIn: 1000,
                 userInfo: {
-                    roleIdentifier: 'roleIdentifier',
+                    roleIdentifier: 'adultRole',
                     serviceId: 'serviceId',
                     email: 'test@example.com',
                     emailVerified: true,
@@ -3183,7 +3197,7 @@ describe('RecordsServer', () => {
                 idToken: 'idToken',
                 expiresIn: 1000,
                 userInfo: {
-                    roleIdentifier: 'roleIdentifier',
+                    roleIdentifier: 'adultRole',
                     serviceId: 'serviceId',
                     email: 'test@example.com',
                     emailVerified: true,
@@ -3293,6 +3307,8 @@ describe('RecordsServer', () => {
                 redirectUri: 'redirectUri',
                 ageOfConsent: 18,
             };
+
+            authController.privoEnabled = true;
         });
 
         it('should return a 200 status code with the registration results', async () => {
@@ -3302,6 +3318,7 @@ describe('RecordsServer', () => {
                 parentServiceId: 'parentServiceId',
                 features: [],
                 updatePasswordLink: 'link',
+                consentUrl: 'consentUrl',
             });
 
             const response = await server.handleHttpRequest(
@@ -3342,6 +3359,7 @@ describe('RecordsServer', () => {
                 parentServiceId: 'parentServiceId',
                 features: [],
                 updatePasswordLink: 'link',
+                consentUrl: 'consentUrl',
             });
 
             const response = await server.handleHttpRequest(
@@ -3388,6 +3406,7 @@ describe('RecordsServer', () => {
                 parentServiceId: 'parentServiceId',
                 features: [],
                 updatePasswordLink: 'link',
+                consentUrl: 'consentUrl',
             });
 
             const response = await server.handleHttpRequest(
@@ -3434,6 +3453,7 @@ describe('RecordsServer', () => {
                 parentServiceId: 'parentServiceId',
                 features: [],
                 updatePasswordLink: 'link',
+                consentUrl: 'consentUrl',
             });
 
             const response = await server.handleHttpRequest(
@@ -3494,6 +3514,7 @@ describe('RecordsServer', () => {
                 parentServiceId: 'parentServiceId',
                 features: [],
                 updatePasswordLink: 'link',
+                consentUrl: 'consentUrl',
             });
 
             const response = await server.handleHttpRequest(
@@ -3547,6 +3568,88 @@ describe('RecordsServer', () => {
                 parentEmail: 'parent@example.com',
                 displayName: 'displayName',
             })
+        );
+    });
+
+    describe('POST /api/v2/privacyFeatures/change', () => {
+        let tenYearsAgo: DateTime;
+
+        beforeEach(async () => {
+            tenYearsAgo = DateTime.now().minus({ years: 10 });
+
+            store.privoConfiguration = {
+                gatewayEndpoint: 'endpoint',
+                featureIds: {
+                    adultPrivoSSO: 'adultAccount',
+                    childPrivoSSO: 'childAccount',
+                    joinAndCollaborate: 'joinAndCollaborate',
+                    publishProjects: 'publish',
+                    projectDevelopment: 'dev',
+                    buildAIEggs: 'buildaieggs',
+                },
+                clientId: 'clientId',
+                clientSecret: 'clientSecret',
+                publicEndpoint: 'publicEndpoint',
+                roleIds: {
+                    child: 'childRole',
+                    adult: 'adultRole',
+                    parent: 'parentRole',
+                },
+                clientTokenScopes: 'scope1 scope2',
+                userTokenScopes: 'scope1 scope2',
+                // verificationIntegration: 'verificationIntegration',
+                // verificationServiceId: 'verificationServiceId',
+                // verificationSiteId: 'verificationSiteId',
+                redirectUri: 'redirectUri',
+                ageOfConsent: 18,
+            };
+
+            const user = await store.findUser(userId);
+            await store.saveUser({
+                ...user,
+                privoServiceId: 'serviceId',
+            });
+
+            privoClientMock.resendConsentRequest.mockResolvedValue({
+                success: true,
+            });
+
+            authController.privoEnabled = true;
+        });
+
+        it('should return a 200 status code', async () => {
+            const response = await server.handleHttpRequest(
+                httpPost(
+                    `/api/v2/privacyFeatures/change`,
+                    JSON.stringify({
+                        userId: userId,
+                    }),
+                    authenticatedHeaders
+                )
+            );
+
+            await expectResponseBodyToEqual(response, {
+                statusCode: 200,
+                body: {
+                    success: true,
+                },
+                headers: accountCorsHeaders,
+            });
+
+            expect(privoClientMock.resendConsentRequest).toHaveBeenCalledWith(
+                'serviceId',
+                'serviceId'
+            );
+        });
+
+        testUrl(
+            'POST',
+            '/api/v2/privacyFeatures/change',
+            () =>
+                JSON.stringify({
+                    userId,
+                }),
+            () => authenticatedHeaders
         );
     });